requirements.txt
flask
easyocr
opencv-python
pytesseract
transformers
torch
scikit-learn
pdf2image
Pillow